Notify commands moved to their own namespace to match project convention
