+++ src/SFA.DAS.ProviderRelationships.MessageHandlers.UnitTests/EventHandlers/ProviderRelationships/DeletedPermissionsEventHandlerTests.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using NUnit.Framework;
 using SFA.DAS.ProviderRelationships.Application.Commands;
+using SFA.DAS.ProviderRelationships.Application.Commands.DeletedPermissionsEventNotify;
 using SFA.DAS.ProviderRelationships.MessageHandlers.EventHandlers.ProviderRelationships;
 using SFA.DAS.ProviderRelationships.Messages.Events;
 using SFA.DAS.ProviderRelationships.ReadStore.Application.Commands.DeletePermissions;
