+++ src/SFA.DAS.ProviderRelationships.MessageHandlers/EventHandlers/ProviderRelationships/DeletedPermissionsEventHandler.cs
@@ -2,6 +2,7 @@ using System.Threading.Tasks;
 using MediatR;
 using NServiceBus;
 using SFA.DAS.ProviderRelationships.Application.Commands;
+using SFA.DAS.ProviderRelationships.Application.Commands.DeletedPermissionsEventNotify;
 using SFA.DAS.ProviderRelationships.Messages.Events;
 using SFA.DAS.ProviderRelationships.ReadStore.Application.Commands.DeletePermissions;
