+++ src/SFA.DAS.ProviderRelationships.MessageHandlers/EventHandlers/ProviderRelationships/UpdatedPermissionsEventHandler.cs
@@ -2,6 +2,7 @@
 using MediatR;
 using NServiceBus;
 using SFA.DAS.ProviderRelationships.Application.Commands;
+using SFA.DAS.ProviderRelationships.Application.Commands.UpdatedPermissionsEventNotify;
 using SFA.DAS.ProviderRelationships.Messages.Events;
 using SFA.DAS.ProviderRelationships.ReadStore.Application.Commands.UpdatePermissions;
