+++ src/SFA.DAS.ProviderRelationships.UnitTests/Application/Commands/DeletedPermissionsEventNotifyCommandHandlerTests.cs
@@ -8,6 +8,7 @@ using NUnit.Framework;
 using SFA.DAS.PAS.Account.Api.Client;
 using SFA.DAS.PAS.Account.Api.Types;
 using SFA.DAS.ProviderRelationships.Application.Commands;
+using SFA.DAS.ProviderRelationships.Application.Commands.DeletedPermissionsEventNotify;
 using SFA.DAS.ProviderRelationships.Data;
 using SFA.DAS.ProviderRelationships.Models;
 using SFA.DAS.ProviderRelationships.UnitTests.Builders;
